refactor: split and structure socket.js file into socket directory
this commit splits the logics originally in the `socket.js` file to
multiple files in the `socket` directory. also, `index.js` now exports
a socket server, not a callback
socket
├── index.js - contains main logic, exports socket server
├── middlewares.js - contains socket middlewares
├── register.js - contains multiple event register functions
└── utils.js - contains (currently user-related) utility functions
